On our server,
queries based on "mail_tracking_event"."tracking_email_id" improved from ~500 ms to ~2 ms
queries based on "mail_tracking_email"."mail_message_id" improved from ~170 ms to ~3 ms

The last ones are run tens of times when a thread has many messages